Added a title which displays time
I wanted to add a title to show the time myself just now. But one can not @lift an observable in a string-interpolation it seems - or at least, I could not figure out how. So I added this to the example, because I think it is a very normal want from an animation, and a less-than-obvious solution.